refactor: /v2/private/wallet/balance
