+++ integrationtest/testdata/feature-inverse-perpetual/v2-private-wallet-balance.json
